Unique ptr
Changed unit tests from shared to unique ptr once I figured out how to
compile. The problem was that initializer lists must be copied, so
unique ptr can not be used in them, as it must be moved. Now unique
pointers are making sure that ownership is clear for the test items.
Fixed missing namespace in logger.
Removed unnecessary constructors.